TicketManagementSystem/requirements.txt
aws-cdk-lib==2.199.0
constructs>=10.0.0,<11.0.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
aws-cdk-lib==2.199.0
aws-cdk.asset-awscli-v1==2.2.237
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
black==25.1.0
boto3==1.38.32
botocore==1.38.32
cattrs==24.1.3
certifi==2025.4.26
cfn-flip==1.3.0
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
comm==0.2.3
constructs==10.4.2
debugpy==1.8.16
decorator==5.2.1
distro==1.9.0
executing==2.2.0
fastjsonschema==2.21.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jiter==0.10.0
jmespath==1.0.1
jsii==1.112.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaleido==1.0.0
langchain-aws==0.2.24
langchain-core==0.3.64
langchain-openai==0.3.21
langsmith==0.3.45
logistro==1.1.0
matplotlib-inline==0.1.7
mypy_extensions==1.1.0
narwhals==2.1.1
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.84.0
orjson==3.10.18
packaging==24.2
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pexpect==4.9.0
platformdirs==4.3.8
plotly==6.3.0
plotly-express==0.4.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
publication==0.0.3
pure_eval==0.2.3
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.27.0
s3transfer==0.13.0
scipy==1.16.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.5
tenacity==9.1.2
tiktoken==0.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==2.13.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
zstandard==0.23.0
